# service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ",
    "version": "0.1.0",
    "description": "Сверка проектной, рабочей и исполнительной документации. Асинхронная pull-модель: загрузка возвращает process_id, клиент опрашивает статус и забирает протокол."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/process/{id}/summary": {
      "get": {
        "summary": "Дашборд объекта: сводка протокола и карта 132 параметров",
        "tags": [
          "process"
        ],
        "description": "Сводка последней версии протокола одним ответом: стадии, воронка по 132 параметрам Матрицы, состояние каждого параметра для карты, разделы, записи по статусам, изменения к прошлой версии. Числа строятся той же функцией, что раздел 2 протокола. Гипотезы свободного поиска в карту Матрицы не входят (ТЗ 9.5), «нет доказательств» — не нарушение (ТЗ 9.3).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "process_id",
                    "object_id",
                    "matrix",
                    "cells"
                  ],
                  "properties": {
                    "process_id": {
                      "type": "string"
                    },
                    "object_id": {
                      "type": "string"
                    },
                    "object_name": {
                      "type": "string"
                    },
                    "protocol": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "scenario_text": {
                      "type": "string",
                      "nullable": true
                    },
                    "stages": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "matrix": {
                      "type": "object",
                      "required": [
                        "total",
                        "compared",
                        "missing_evidence",
                        "not_comparable",
                        "not_checked"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "compared": {
                          "type": "integer"
                        },
                        "with_violation": {
                          "type": "integer"
                        },
                        "with_candidate": {
                          "type": "integer"
                        },
                        "missing_evidence": {
                          "type": "integer"
                        },
                        "not_comparable": {
                          "type": "integer"
                        },
                        "not_checked": {
                          "type": "integer"
                        },
                        "not_applicable": {
                          "type": "integer"
                        },
                        "out_of_scope": {
                          "type": "integer"
                        },
                        "hypothesis": {
                          "type": "integer"
                        }
                      }
                    },
                    "records": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "integer"
                      }
                    },
                    "sections": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "changes": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "added": {
                          "type": "integer"
                        },
                        "changed": {
                          "type": "integer"
                        },
                        "removed": {
                          "type": "integer"
                        },
                        "decided_changed": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "sync": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "status": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "version": {
                            "type": "integer"
                          },
                          "created_at": {
                            "type": "string",
                            "nullable": true
                          },
                          "records": {
                            "type": "integer"
                          },
                          "candidates": {
                            "type": "integer"
                          },
                          "compared": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "cells": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "code",
                          "state"
                        ],
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "id": {
                            "type": "integer",
                            "nullable": true
                          },
                          "name": {
                            "type": "string",
                            "nullable": true
                          },
                          "section": {
                            "type": "string",
                            "nullable": true
                          },
                          "queue": {
                            "type": "integer",
                            "nullable": true
                          },
                          "critical": {
                            "type": "boolean",
                            "nullable": true
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "VIOLATION",
                              "CANDIDATE",
                              "CLEAN",
                              "NOT_COMPARABLE",
                              "MISSING_EVIDENCE",
                              "NOT_APPLICABLE",
                              "NOT_CHECKED",
                              "OUT_OF_SCOPE",
                              "HYPOTHESIS"
                            ]
                          },
                          "group": {
                            "type": "string"
                          },
                          "records": {
                            "type": "integer"
                          },
                          "reason": {
                            "type": "string",
                            "nullable": true
                          },
                          "note": {
                            "type": "string",
                            "nullable": true
                          },
                          "finding": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/{id}/report": {
      "get": {
        "summary": "Протокол проверки объекта (JSON)",
        "tags": [
          "process"
        ],
        "description": "Документ для инспектора по образцу Приложения 2 к ТЗ: статус загрузки и тип проверки, сводка по 132 параметрам Матрицы, пять таблиц (комплектность и сопоставимость, кандидаты по уровню риска, подтверждённые, проверенные отрицательные, гипотезы), резолютивная часть после финализации, карточки доказательств с SHA-256, редакциями и решениями, правила подсчёта. Решения инспектора берутся на момент запроса. Файл сдачи организатору — отдельно, GET /api/v1/process/{id}/protocol.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "document",
                    "object",
                    "protocol",
                    "traceability",
                    "section_1_upload",
                    "section_2_summary",
                    "tables",
                    "evidence_cards"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "document": {
                      "type": "string"
                    },
                    "number": {
                      "type": "string",
                      "nullable": true
                    },
                    "object": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "protocol": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "traceability": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "section_1_upload": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "section_2_summary": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "tables": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "evidence_cards": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/{id}/report.pdf": {
      "get": {
        "summary": "Протокол проверки объекта (PDF)",
        "tags": [
          "process"
        ],
        "description": "Документ для инспектора по образцу Приложения 2 к ТЗ: статус загрузки и тип проверки, сводка по 132 параметрам Матрицы, пять таблиц (комплектность и сопоставимость, кандидаты по уровню риска, подтверждённые, проверенные отрицательные, гипотезы), резолютивная часть после финализации, карточки доказательств с SHA-256, редакциями и решениями, правила подсчёта. Решения инспектора берутся на момент запроса. Файл сдачи организатору — отдельно, GET /api/v1/process/{id}/protocol.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "PDF протокола",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process/{id}/report.docx": {
      "get": {
        "summary": "Протокол проверки объекта (DOCX)",
        "tags": [
          "process"
        ],
        "description": "Документ для инспектора по образцу Приложения 2 к ТЗ: статус загрузки и тип проверки, сводка по 132 параметрам Матрицы, пять таблиц (комплектность и сопоставимость, кандидаты по уровню риска, подтверждённые, проверенные отрицательные, гипотезы), резолютивная часть после финализации, карточки доказательств с SHA-256, редакциями и решениями, правила подсчёта. Решения инспектора берутся на момент запроса. Файл сдачи организатору — отдельно, GET /api/v1/process/{id}/protocol. DOCX — то же содержание, что PDF (оба формата собираются из одних блоков), в виде, который открывается и правится в Word: ТЗ, модуль 7, «экспорт протоколов в PDF, DOCX, XML».",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "DOCX протокола",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}